be gio hang chi tiet

+++ back-end/src/main/java/com/poly/application/service/impl/GioHangChiTietServiceImpl.java
@@ -32,7 +32,13 @@ public class GioHangChiTietServiceImpl implements GioHangChiTietService {
 
     @Override
     public void delete(Long idGioHangChiTiet) {
-        GioHangChiTiet gioHangChiTiet= repository.tim(idGioHangChiTiet);
+        Optional<GioHangChiTiet> optional = repository.findById(idGioHangChiTiet);
+
+        if (optional.isEmpty()) {
+            throw new NotFoundException("Giỏ hàng chi tiết không tồn tại");
+        }
+
+        GioHangChiTiet gioHangChiTiet = optional.get();
         repository.delete(gioHangChiTiet);
     }
 
@@ -44,14 +50,8 @@ public class GioHangChiTietServiceImpl implements GioHangChiTietService {
         }
 
         GioHangChiTiet gioHangChiTiet = optional.get();
+        gioHangChiTiet.setSoLuong(request.getSoLuong());
         mapper.convertUpdateRequestToEntity(request, gioHangChiTiet);
         return mapper.convertEntityToResponse(repository.save(gioHangChiTiet));
     }
-
-    @Override
-    public GioHangChiTietResponse getOne(Long idGioHangChiTiet) {
-        GioHangChiTiet gioHangChiTiet = repository.tim(idGioHangChiTiet);
-        return mapper.convertEntityToResponse(gioHangChiTiet);
-    }
-
 }
